Change gosh to build the program in a module-aware directory
The gosh-generated file used to be just a single file but now it will
generate a directory and write the file in there. Also, if the
GO111MODULE environment variable is not set to off it will generate the
go.mod and go.sum files as needed.

Also:
- Add lots of verbose messages about what gosh is doing and how long various
functions take to run.
- Expand the program description
- Quieten the programs called (unless an error is detected)
- Use the latest versions of packages