Add support for snippets to the gosh program
Also:

Use the latest versions of packages.

Use the filecheck standard Provisos functions

Replace the begin and end parameters with before and after, this regularises
the parameter names.

Allow the gosh temp directory to be set through the command line

Add a config file to gosh - this is useful for the temp and snippets
directories.

Refactor the scripts so preScript becomes beforeScript and postScript becomes
afterScript.