Fixed the Regexp used to check valid Go names
This now allows private types and is properly anchored at the start of the
name. Also expanded the help text for the 'type-name' parameter.

Also refactored the code generation a bit to have separate functions for the
type declaration and constants and for the IsValid func

Removed the sentinel values and changed IsValid correspondingly