Process additional args as non-files if no readloop params are given
If additional arguments are given but no readloop parameters are given then
loop over the arguments but don't try to open and read them. It's no longer
an error to give trailing parameters (after '--') and not to give some
readloop argument.

Also
- we don't change filename arguments into absolute paths (we don't need
to since we now chdir back to the directory where we ran gosh from).
- we only check for corresponding .orig files if we are editing in place
- there are two new 'known' variables: _arg and _args