Substantial refactoring to support snippet order checking
The Gosh scripts (global, before etc) are now all held in a single map of
string to slice of ScriptEntries.

A ScriptEntry is a structure which holds a key and an expansion function. The
key can be the name of a snippet which is held in the snippetCache. The
executable text of a snippet is no longer just written into the script as
each argument is parsed but now it is cached as the argument is parsed and
then expanded into the script as the program is written. This allows for
correct checking of the order of snippets rather than just checking that the
arguments were given in the 'correct' order.

Also the filesErrorMap in the Gosh structure is renamed to errMap and is used
for multiple error reporting. The corresponding AddErr method is now called
addError and is used by a number of post-argument-parsing error reports.

snippets are now held as a struct rather than just being pasted into the
script directly. They are cached on the first occurence.