Change the snippet directory note to not list the directories
The list of snippet directories is available but it is user specific as one
of the snippet directories is in the .config directory in $HOME and so when
we take the notes and add them to the markdown files they are
incorrect. Consequently we give instead the parameter name for the user to
find out the directories.