Add the build-arg parameter to mkdoc
This allows the build to be configured. In particular it allows us to build
executables having version information to be built with the initial check
that the version details are present turned off.

Also the built program is now written to a temporary file rather than
overwriting the file in the source directory (if any). This temporary file
will be automatically removed when mkdoc exits.