Force the import of necessary oackages that we know we'll need
This reduces the reliance on the goimports command. Specifically if any of
the print parameters are given we add 'fmt' and we always add 'os' because
the first thing the generated program calls is 'os/Chdir(...)' to get itself
back from the temporary build directory to where it was invoked.