Refactor the mkdoc program
Provide head and tail files for each of the sections

Make the generation of the Markdown table driven to ensure consistency and
reduce repetitive code.

Take snippets of Markdown from modules used by the program to provide
consistent documentation across programs. The extra parameters
'snippet-mod-prefix' and 'snippet-mod-skip' are added to control this
behaviour.