Add support for repeat editing
There is a new 'edit-repeat' parameter and the code has been refactored to
allow a loop where the editor is called, the program is built and run and the
user is prompted to see if they want to edit/build/run again.

Also changed the behaviour of the edit parameter - it no longer forces the
code to be kept.

Also several places where errors were reported and the code exits didn't used
to tell the user the gosh directory name; this is now fixed. Also some error
messages didn't have a newline at the end; this is also fixed.