Add the internal packages callstack and stdparams
callstack is used to report timings and to give tags to debug (verbose)
messages that reflect the call stack

stdparams is used to hold common parameters to ensure that they use the same
names across programs. Currently only the AddTiming func, used to add the
show-timings parameter exists.

These packages may "graduate" in the future out of the internal playpen but
for now it allows some experimentation without needing to support the
packages and freeze the API.