refactor(gosh): check if the in-place edit params are inconsistent
We report an error if the print options for writing to the in-place edit file
are being used but we are not doing an in-place edit. Note that as this is a
desctructive operation (it will change a file) we do not simply set the
in-place edit flag when we see such a parameter.