feature(gosh): add support for named imports
The user can now give an import name with an embedded '=' and the import
statement generated will use the text before the '=' as an import name. You
can also give a name of '.' which will place the package API in the local
namespace (so you don't need to use the package name).