feature(gosh): use gopls as a formatter if available
Note that we use gopls with parameter 'imports' so that it will populate the
import statement in the generated code. We do not use it to format the code
so this is strictly a misnomer.

We now have a list of alternative formatters rather than defaulting to gofmt
and using goimports if available.

Also change the notes to no longer refer to goimports directly but now as
just one of several 'formatters' that will be used if available.

The ...file params now all refer to each other in the See-Also
section. Similarly for the script-editing and formatting parameters.