fix(gosh): call the formatter with the filename not the whole path
The bug being fixed is that "gopls imports -w <filename>" on MacOS will not
update the import statement correctly  if filename is the full path to the
temp directory but will if it is just the filename. This is safe to change
because the formatting and file writing are done after changing directory to
the temp directory.

Also change the name of the Gosh member from filename to filepath and add a
constant goshFilename for the gosh.go file.