feature(gosh): separate the formatting from the import generation
The automatic population of the import statements used to be done as a side
effect of the formatting of the program. Now these two steps are done
separately. The formatting stage is off by default and must be explicitly
performed by passing the 'format' parameter. The import statement populating
is on by default.