feature(gosh): allow the environment for the program to be set
The new 'env' parameter now allows gosh to override environment values when
the generated program runs. Existing values can be changed or new values
set. If the 'clear-env' parameter is given then the existing environment is
cleared.

The '_' env var is always set to the full path of the generated program.