fix(mkparamfilefunc): correct the group file generation
A recent change (90802a59f0b6f59e029f8943641f63632c9015b3 - 2023-Jun-11)
changed the behaviour of the makeGroupSuffix func so that it changed the
prog.groupName value rather than a local copy. This has been corrected.

Ther effect of this bug was to change the name of the parameter group and
also of the parameter file.