fix(findCmpRm): order the command Starts and Waits correctly
There was a bug where if the files being compared were large the diff command
hadn't finished writing its output when the less command quit. This caused
the Wait on the diffCmd to never return because it was blocking on the write
to the pipe that was never cleared because less had exited. All very
embarassing; as if I'd never written code before.